+++ docs/lib/examples/NavVertical.js
@@ -6,7 +6,7 @@ export default class Example extends React.Component {
     return (
       <div>
         <p>List Based</p>
-        <Nav inline>
+        <Nav vertical>
           <NavItem>
             <NavLink href="#">Link</NavLink>
           </NavItem>
@@ -22,7 +22,7 @@ export default class Example extends React.Component {
         </Nav>
         <hr />
         <p>Link based</p>
-        <Nav inline>
+        <Nav vertical>
           <NavLink href="#">Link</NavLink> <NavLink href="#">Link</NavLink> <NavLink href="#">Another Link</NavLink> <NavLink disabled href="#">Disabled Link</NavLink>
         </Nav>
       </div>