+++ src/Modal.js
@@ -1,4 +1,4 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import classNames from 'classnames';
 import omit from 'lodash.omit';
@@ -11,6 +11,7 @@ import {
   mapToCssModules,
 } from './utils';
 
+const { PropTypes } = React;
 const propTypes = {
   isOpen: PropTypes.bool,
   size: PropTypes.string,
@@ -39,7 +40,7 @@ const defaultProps = {
   isOpen: false,
   backdrop: true,
   keyboard: true,
-  zIndex: 1000,
+  zIndex: 1050,
 };
 
 class Modal extends React.Component {
@@ -191,6 +192,7 @@ class Modal extends React.Component {
             transitionLeaveTimeout={300}
             onClickCapture={this.handleBackdropClick}
             onKeyUp={this.handleEscape}
+            cssModule={cssModule}
             className={mapToCssModules(classNames('modal', modalClassName), cssModule)}
             style={{ display: 'block' }}
             tabIndex="-1"
@@ -215,6 +217,7 @@ class Modal extends React.Component {
             transitionAppearTimeout={150}
             transitionEnterTimeout={150}
             transitionLeaveTimeout={150}
+            cssModule={cssModule}
             className={mapToCssModules(classNames('modal-backdrop', backdropClassName), cssModule)}
           />
         )}