+++ src/Nav.js
@@ -1,13 +1,12 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import classNames from 'classnames';
 import { mapToCssModules } from './utils';
 
+const { PropTypes } = React;
 const propTypes = {
-  inline: PropTypes.bool,
-  disabled: PropTypes.bool,
   tabs: PropTypes.bool,
   pills: PropTypes.bool,
-  stacked: PropTypes.bool,
+  vertical: PropTypes.bool,
   navbar: PropTypes.bool,
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
   className: PropTypes.string,
@@ -24,8 +23,7 @@ const Nav = (props) => {
     cssModule,
     tabs,
     pills,
-    inline,
-    stacked,
+    vertical,
     navbar,
     tag: Tag,
     ...attributes
@@ -37,9 +35,7 @@ const Nav = (props) => {
     {
       'nav-tabs': tabs,
       'nav-pills': pills,
-      'nav-inline': inline,
-      'nav-stacked': stacked,
-      disabled: attributes.disabled
+      'flex-column': vertical
     }
   ), cssModule);