+++ src/Tooltip.js
@@ -1,12 +1,16 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import classNames from 'classnames';
 import omit from 'lodash.omit';
 import TetherContent from './TetherContent';
 import { getTetherAttachments, tetherAttachements, mapToCssModules } from './utils';
 
+const { PropTypes } = React;
 const propTypes = {
   placement: React.PropTypes.oneOf(tetherAttachements),
-  target: PropTypes.string.isRequired,
+  target: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.object
+  ]).isRequired,
   isOpen: PropTypes.bool,
   disabled: PropTypes.bool,
   tether: PropTypes.object,
@@ -51,6 +55,7 @@ class Tooltip extends React.Component {
     super(props);
 
     this.addTargetEvents = this.addTargetEvents.bind(this);
+    this.getTarget = this.getTarget.bind(this);
     this.getTetherConfig = this.getTetherConfig.bind(this);
     this.handleDocumentClick = this.handleDocumentClick.bind(this);
     this.removeTargetEvents = this.removeTargetEvents.bind(this);
@@ -64,7 +69,7 @@ class Tooltip extends React.Component {
   }
 
   componentDidMount() {
-    this._target = document.getElementById(this.props.target);
+    this._target = this.getTarget();
     this.addTargetEvents();
   }
 
@@ -113,12 +118,20 @@ class Tooltip extends React.Component {
     return delay;
   }
 
+  getTarget() {
+    const { target } = this.props;
+    if (typeof target === 'object') {
+      return target;
+    }
+    return document.getElementById(target);
+  }
+
   getTetherConfig() {
     const attachments = getTetherAttachments(this.props.placement);
     return {
       ...defaultTetherConfig,
       ...attachments,
-      target: '#' + this.props.target,
+      target: this.getTarget,
       ...this.props.tether
     };
   }