+++ src/__tests__/DropdownItem.spec.js
@@ -19,6 +19,41 @@ describe('DropdownItem', () => {
     expect(wrapper.find('button').length).toBe(1);
   });
 
+  it('should render type as "button" by default when tag is "button"', () => {
+    const wrapper = mount(<DropdownItem>Home</DropdownItem>);
+
+    expect(wrapper.find('button').prop('type')).toBe('button');
+    expect(wrapper.text()).toBe('Home');
+  });
+
+  it('should render type as undefined by default when tag is "button" and toggle is false', () => {
+    const wrapper = mount(<DropdownItem toggle={false}>Home</DropdownItem>);
+
+    expect(wrapper.find('button').prop('type')).toBe(undefined);
+    expect(wrapper.text()).toBe('Home');
+  });
+
+  it('should render type as "button" by default when tag is "button" and onClick is provided', () => {
+    const wrapper = mount(<DropdownItem onClick={() => {}}>Home</DropdownItem>);
+
+    expect(wrapper.find('button').prop('type')).toBe('button');
+    expect(wrapper.text()).toBe('Home');
+  });
+
+  it('should render type as user defined when defined by the user', () => {
+    const wrapper = mount(<DropdownItem type="submit">Home</DropdownItem>);
+
+    expect(wrapper.find('button').prop('type')).toBe('submit');
+    expect(wrapper.text()).toBe('Home');
+  });
+
+  it('should not render type by default when the type is not defined and the tag is not "button"', () => {
+    const wrapper = mount(<DropdownItem tag="a">Home</DropdownItem>);
+
+    expect(wrapper.find('a').prop('type')).toBe(undefined);
+    expect(wrapper.text()).toBe('Home');
+  });
+
   it('should render custom element', () => {
     const Link = (props) => <a href="/home" {...props}>{props.children}</a>;
     const wrapper = mount(<DropdownItem tag={Link}>Home</DropdownItem>);
@@ -43,6 +78,16 @@ describe('DropdownItem', () => {
       const wrapper = mount(<DropdownItem divider />);
 
       expect(wrapper.find('.dropdown-divider').length).toBe(1);
+    });
+  });
+
+  describe('link (with href)', () => {
+    it('should render an anchor tag', () => {
+      const wrapper = mount(<DropdownItem href="#">GO!</DropdownItem>);
+
+      expect(wrapper.find('a').length).toBe(1);
+      expect(wrapper.find('a').hasClass('dropdown-item')).toBe(true);
+      expect(wrapper.text()).toBe('GO!');
     });
   });